Slice header size
The ParseSliceHeader function responds with a slice header and a bytesRead value. This can be the slice header size in Bytes.

The SPS file populates a `NrBytesRead` value in the `SPS` structure.
Haven't used that method as I feel it's a better approach to be keep the Size and the Params as separate.