Make boxHeader and fields public
`boxHeader` is used in `BoxDecoder` functions, however, this struct and its
fields were never public. This changes the struct and fields to be
public as well as the `decodeHeader` function.

This might seem like a breaking change for codebases using mp4ff.
However, since `boxHeader` was never public `BoxDecoder` functions could
not be called directly (appart from passing `nil`).